♻️ refactor: 클래스명 수정(adapter -> helper)
