+++ sycls-admin-api/src/main/java/com/b1/seat/SeatGradeHelper.java
@@ -6,10 +6,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
 
-@Slf4j(topic = "Seat Grade Adapter")
+@Slf4j(topic = "Seat Grade Helper")
 @Repository
 @RequiredArgsConstructor
-public class SeatGradeAdapter {
+public class SeatGradeHelper {
 
     private final SeatGradeRepository seatGradeRepository;
